requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Werkzeug==2.2.2
mysql
mysql-connector-python==8.0.15
python-dotenv
